spike(rendering): factor out MutableChunkMesh
