+++ engine/src/main/java/org/terasology/engine/monitoring/chunk/ChunkMeshInfo.java
@@ -1,8 +1,9 @@
-// Copyright 2021 The Terasology Foundation
+// Copyright 2022 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 package org.terasology.engine.monitoring.chunk;
 
 import org.terasology.engine.rendering.primitives.ChunkMesh;
+import org.terasology.engine.rendering.primitives.MutableChunkMesh;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -14,7 +15,7 @@ public class ChunkMeshInfo {
     public final int totalTimeToGenerateBlockVertices;
     public final int totalTimeToGenerateOptimizedBuffers;
 
-    public ChunkMeshInfo(ChunkMesh mesh) {
+    public ChunkMeshInfo(MutableChunkMesh mesh) {
         checkNotNull(mesh, "The parameter 'mesh' must not be null");
 
         int vertices = 0;
@@ -22,7 +23,7 @@ public class ChunkMeshInfo {
 
         if (mesh.hasVertexElements()) {
             for (ChunkMesh.RenderType type : ChunkMesh.RenderType.values()) {
-                final ChunkMesh.VertexElements element = mesh.getVertexElements(type);
+                final MutableChunkMesh.VertexElements element = mesh.getVertexElements(type);
                 vertices += element.buffer.elements();
                 indices += element.indices.indices();
             }
