+++ engine/src/main/java/org/terasology/engine/persistence/typeHandling/extensionTypes/ChunkMeshTypeHandler.java
@@ -1,4 +1,4 @@
-// Copyright 2021 The Terasology Foundation
+// Copyright 2022 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
 package org.terasology.engine.persistence.typeHandling.extensionTypes;
@@ -6,6 +6,7 @@ package org.terasology.engine.persistence.typeHandling.extensionTypes;
 import org.lwjgl.BufferUtils;
 import org.terasology.engine.rendering.primitives.ChunkMesh;
 import org.terasology.engine.rendering.primitives.ChunkMeshImpl;
+import org.terasology.engine.rendering.primitives.MutableChunkMesh;
 import org.terasology.persistence.typeHandling.PersistedData;
 import org.terasology.persistence.typeHandling.PersistedDataSerializer;
 import org.terasology.persistence.typeHandling.TypeHandler;
@@ -15,9 +16,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class ChunkMeshTypeHandler extends TypeHandler<ChunkMesh> {
+public class ChunkMeshTypeHandler extends TypeHandler<MutableChunkMesh> {
     @Override
-    protected PersistedData serializeNonNull(ChunkMesh value, PersistedDataSerializer serializer) {
+    protected PersistedData serializeNonNull(MutableChunkMesh value, PersistedDataSerializer serializer) {
         if (!value.hasVertexElements()) {
             throw new IllegalStateException("Attempting to serialize a ChunkMesh whose data has already been discarded.");
         }
@@ -35,7 +36,7 @@ public class ChunkMeshTypeHandler extends TypeHandler<ChunkMesh> {
     }
 
     @Override
-    public Optional<ChunkMesh> deserialize(PersistedData data) {
+    public Optional<MutableChunkMesh> deserialize(PersistedData data) {
         List<ByteBuffer> asBuffers = new ArrayList<>();
         for (PersistedData datum : data.getAsArray()) {
             ByteBuffer buffer = datum.getAsByteBuffer();
@@ -44,7 +45,7 @@ public class ChunkMeshTypeHandler extends TypeHandler<ChunkMesh> {
             directBuffer.rewind();
             asBuffers.add(directBuffer);
         }
-        ChunkMesh result = new ChunkMeshImpl();
+        MutableChunkMesh result = new ChunkMeshImpl();
         for (ChunkMesh.RenderType renderType : ChunkMesh.RenderType.values()) {
             result.getVertexElements(renderType).buffer.replace(asBuffers.remove(0));
             result.getVertexElements(renderType).indices.replace(asBuffers.remove(0));
