+++ engine/src/main/java/org/terasology/engine/rendering/logic/ChunkMeshComponent.java
@@ -4,7 +4,7 @@
 package org.terasology.engine.rendering.logic;
 
 import org.terasology.engine.network.Replicate;
-import org.terasology.engine.rendering.primitives.ChunkMesh;
+import org.terasology.engine.rendering.primitives.MutableChunkMesh;
 import org.terasology.joml.geom.AABBf;
 
 /**
@@ -13,7 +13,7 @@ import org.terasology.joml.geom.AABBf;
  */
 public class ChunkMeshComponent implements VisualComponent<ChunkMeshComponent> {
     @Replicate
-    public ChunkMesh mesh;
+    public MutableChunkMesh mesh;
     @Replicate
     public AABBf aabb;
     @Replicate
@@ -21,12 +21,12 @@ public class ChunkMeshComponent implements VisualComponent<ChunkMeshComponent> {
 
     public ChunkMeshComponent() { }
 
-    public ChunkMeshComponent(ChunkMesh mesh, AABBf aabb) {
+    public ChunkMeshComponent(MutableChunkMesh mesh, AABBf aabb) {
         this.mesh = mesh;
         this.aabb = aabb;
     }
 
-    public synchronized void setMesh(ChunkMesh mesh) {
+    public synchronized void setMesh(MutableChunkMesh mesh) {
         if (this.mesh != null) {
             this.mesh.dispose();
         }
