+++ engine/src/main/java/org/terasology/engine/world/chunks/RenderableChunk.java
@@ -1,11 +1,11 @@
-// Copyright 2021 The Terasology Foundation
+// Copyright 2022 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 package org.terasology.engine.world.chunks;
 
 import org.joml.Vector3f;
-import org.terasology.joml.geom.AABBfc;
-import org.terasology.gestalt.module.sandbox.API;
 import org.terasology.engine.rendering.primitives.ChunkMesh;
+import org.terasology.gestalt.module.sandbox.API;
+import org.terasology.joml.geom.AABBfc;
 
 /**
  * Anything that acts like a chunk for rendering purposes
